requirements.txt
Authlib==1.3.1
Flask==3.0.3
Flask-Compress==1.15
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.32
